firewire: ohci: facilitate regAccessFail debugging
I experimented with making the IsochronousCycleTimer failures a soft
error and this has the driver working for me without having to probe it
while the cable is connected.

The offending reg_rw_sclk() call is around line 2439 of ohci.c.  I
attached a patch that enables this feature when you load the module with
cycle_timer_hard_fail=0 as a parameter as a proof of concept.

Would it make sense to introduce a workaround for this as a quirk or
parameter?