firewire: ohci: handle register access failure in SClk domain
One of the changes from OHCI-1394 v1.0 to v1.1 is that the PHY's SClk
signal may sometimes not be present during normal operation, and
accesses to certain registers fail then.  See OHCI-1394 v1.1 sections
1.4.1, 4., and 6.1.  The specification does not tell us though how to
recover from this condition.

So far, regAccessFail has been seen only with the following devices:

  - Texas Instruments PCIxx21 FireWire + CardBus + flash memory card
    controller in a Toshiba Satellite:
    https://bugzilla.redhat.com/show_bug.cgi?id=608544

  - O2 Micro FireWire + flash memory card controller in various Dell
    laptops:
    https://bugs.launchpad.net/ubuntu/+source/linux/+bug/801719
    https://bugs.launchpad.net/ubuntu/+source/linux/+bug/881688
    http://marc.info/?l=linux1394-devel&m=132309283531423
    http://marc.info/?l=linux1394-devel&m=132368567907469
    https://bugzilla.kernel.org/show_bug.cgi?id=43247
    and several more reports.

  - Pinnacle MovieBoard:
    http://git.kernel.org/?p=linux/kernel/git/torvalds/linux.git;a=commit;h=7f7e37115a8b
    http://marc.info/?l=linux1394-devel&m=130714243325962
    In case of the Movieboard, the regAccessFail sightings are
    misleading:  This controller simply starts to throw random corrupt
    IRQ event flags at some point.

This patch adds a check for regAccessFail at each and every register
access within the SClk domain.  If the access failure is encountered,
an error is propagated to upper layers.  Tests with a prior version of
this patch on an O2Micro controller have shown that retry loops do not
get us out of a regAccessFail situation.

Signed-off-by: Stefan Richter <stefanr@s5r6.in-berlin.de>